Feat: implement rectangle area and perimeter functions

+++ src/area/circle.h
@@ -4,12 +4,12 @@
     #define M_PI 3.14159265358979323846
 #endif
 
-static inline double area_circle(double radius) 
+static inline float area_circle(float radius) 
 {
     return M_PI * (radius * radius);
 }
 
-static inline double circumference_circle(double radius) 
+static inline float circumference_circle(float radius) 
 {
     return 2.0 * M_PI * radius;
 }
